% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtpreparation.R
\name{add_id_column}
\alias{add_id_column}
\title{Helper function to add ID column}
\usage{
add_id_column(data, keyword, category)
}
\arguments{
\item{data}{A tibble or dataframe containing a Google Trends
time series.}

\item{keyword}{The keyword or keywords that have been searched in
the prior function.}

\item{category}{The category ID from Google Trends.}
}
\value{
A tibble with an added ID column, if there was no beforehand.
Otherwise, the original data will be returned.
}
\description{
This function checks if the dataframe from Google
contains a column "id". If there is no said column,
this function adds one with the fitting ID from Google.
}
\keyword{internal}
