% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_series.R
\name{daily_series}
\alias{daily_series}
\title{Generate a consistent daily time series}
\usage{
daily_series(keyword = c("arbeitslos"), geo = "DE", from = "2006-01-01")
}
\arguments{
\item{keyword}{A character vector consisting of the
search query. As of now, you can only enter one single
keyword. Categories are not possible.}

\item{geo}{A geographical region to restrict the search query to.}

\item{from}{Start date of the search query.}
}
\value{
Tibble with daily relative search volumes.
}
\description{
Since Google only provides data for a short time
period of less than nine months, one needs to apply tricks
to get a longer time series with daily data.
\code{daily_series} estimates
via the Chow-Lin method a consistent long daily time series based
on monthly data.
}
\section{Warning}{

This function takes a long time and generates a lot of
queries at Google. An IP ban is therefore quite likely.
}

\examples{
\dontrun{
daily_series(keyword = "Ikea", geo = "NL", from = "2021-01-01")
}
}
