% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trend.R
\name{est_trend}
\alias{est_trend}
\title{Estimates the trend of Google search queries}
\usage{
est_trend()
}
\value{
Returns a time series which consists of a polynomial trend
of degree five of a category.
}
\description{
\code{est_trend} downloads for a sample
of 250 random Google Trends categories the relative search volume.
Based on this, a common trend is calculated.
}
\examples{
\dontrun{
est_trend()
}
}
