% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorR2.R
\name{factorR2}
\alias{factorR2}
\title{\eqn{R^2} der Regression der Serien auf die Faktoren}
\usage{
factorR2(series, factors, plot = F)
}
\arguments{
\item{series}{tibble mit den Zeitreihen als Spalten.}

\item{factors}{tibble mit den Faktoren als Spalten.}

\item{plot}{plot=TRUE gibt zusaetzlich einen Barplot der \eqn{R^2} aus.}
}
\value{
Tabelle der \eqn{R^2} jeder Zeitreihe auf jeden Faktor. Wenn plot=T, dann wird
zusaetzlich ein Plot zurueckgegeben.
}
\description{
\code{factorR2} gibt \eqn{R^2} der Regression der Serien auf die Faktoren als Tabelle aus.
}
\examples{
\dontrun{
dat <- pca(keywords = c("ikea", "saturn", "amazon", "ebay"),
categories = 0, geo = "DE", start = "2006-01-01", end = Sys.Date(),
components = max(length(keywords), length(categories)))

series <- dat \%>\% select(date, 6:9)
factors <- dat \%>\% select(date, 2:5)

factorR2(series, factors, plot = T)
}
}
