% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_index.R
\name{g_index}
\alias{g_index}
\title{Aufbereitung der Google Daten.}
\usage{
g_index(
  keyword = NA,
  category = 0,
  geo = "DE",
  time = str_c("2006-01-01 ", Sys.Date())
)
}
\arguments{
\item{keyword}{Eine character-Vektor mit dem Suchbegriffen}

\item{category}{Ein Numeric Vektor mit den Kategorien}

\item{geo}{Die Region}

\item{time}{Der Zeitraum in der Angabe wie in ts_gtrends}
}
\value{
Für jede Reihe wird zunächst eine log-Trafo durchgeführt.
Dann wird mit JDemetra eine Saisonbereinigung mit X-13 Arima durchgeführt.
Dann werden erste Differenzen zurückgegeben. Standardmäßig mit lag
}
\description{
\code{g_index} lädt für mehrere Suchbegriffe/Kategorien Daten herunter, führt log Trafo und Saisonbereinigung durch und gibt (gelagte) erste Differenzen aus
}
