% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtpreparation.R
\name{gtpreparation}
\alias{gtpreparation}
\title{Preparation of Google Trends data}
\usage{
gtpreparation(
  keyword = NA,
  category = 0,
  geo = "DE",
  time = str_c("2006-01-01 ", Sys.Date()),
  lags = 0
)
}
\arguments{
\item{keyword}{A vector (chr) with search requests.}

\item{category}{A category ID from Google Trends. As of now,
only one category can be given. If you need to use more categories,
the use of a \code{for}-loop is recommended.}

\item{geo}{A geographical region to restrict the search queries to.}

\item{time}{Time period from where the relative values should be taken
(for more information, visit the
documentation of \code{\link[trendecon]{ts_gtrends}}). Attention:
As this function will only work on monthly data, you need to enter
for time a time frame that is longer than 5 years.
Otherwise, the time series from Google Trends will be based on
monthly or daily data and cannot be evaluated.}

\item{lags}{Number of delays in additional columns (max. value: 4).}
}
\value{
Firstly, each row will be log transformed and
seasonal adjusted (via \code{seasonal}'s X-13 ARIMA methods).
Furthermore, the first derivatives of these adjusted time
series will be returned (optionally with additional columns containing
lags).
}
\description{
\code{gtpreparation} downloads for
various search requests respective trends categories
data and applies a logarithmic transformation aswell as a
seasonal adjustment.
The function returns seasonal adjusted
first derivatives (lagged if desired).
}
\examples{
gtpreparation(keyword = "ikea", time = "2015-01-01 2021-01-01")
}
