% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtseas_adj.R
\name{gtseas_adj}
\alias{gtseas_adj}
\title{Seasonal adjustment of a Google Trends time series}
\usage{
gtseas_adj(
  timeseries = NULL,
  keyword = NA,
  category = NA,
  geo = "DE",
  timeframe = paste("2006-01-01", Sys.Date()),
  method = "arima",
  freq = "monthly",
  log.trafo = FALSE
)
}
\arguments{
\item{timeseries}{A already created time
series to seasonal adjust. Preferably already as tsibble or tibble.
If another data format is given, it will be coerced to a tsibble.}

\item{keyword}{A vector (chr) of keywords to search for.}

\item{category}{A vector (numeric) of category numbers to search for.}

\item{geo}{The region to search in.}

\item{timeframe}{A time frame to search the queries in consisting of
a start date and an end date in YYYY-MM-DD form.}

\item{method}{Character, which method for
adjustment should be choosen. See Details.}

\item{freq}{Character "month" or "quarter" for the frequency.}

\item{log.trafo}{Logical,
indicates if value should be transformed to log(value).}
}
\value{
Returns a tsibble with sesonal adjusted values and a
date column. Any key column will be lost. Therefore, you should only
to this with single time series with one keyword or category.
}
\description{
This function can either seasonal adjust a given time series
or create a new time series based on a Google Trends search query
and then directly seasonally adjust it.
}
\section{Methods}{

For method, there can be choosen \code{"firstdiff"} and \code{"arima"}.
If \code{"firstdiff"}, first derivatives with \code{lag = 1} are computed.
If \code{"arima"}, the X-13ARIMA-SEATS  procedure is used
(performed by \code{\link[seasonal:seas]{seasonal::seas()}} from the seasonal package).
}

\examples{
series <- trendecon::ts_gtrends(c("ikea", "saturn"), time = "2020-01-01 2021-06-01")
gtseas_adj(series, freq = "month", log.traf = TRUE, method = "firstdiff")

gtseas_adj(category = 179, timeframe = "2015-01-01 2021-01-01", method = "arima")

}
