% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_adj.R
\name{helper_adj}
\alias{helper_adj}
\title{Basic adjustment via helper function}
\usage{
helper_adj(series, log.trafo = F)
}
\arguments{
\item{series}{The input tibble in tidy form
with columns \code{time}, \code{value} and optional column \code{id}.}

\item{log.trafo}{Logical, indicates if value
should be transformed to log(value).}
}
\value{
This function returns a tibble.
}
\description{
This function is a helper function to reduce
duplicate code in\code{\link{trend_adj}()}
and \code{\link{seas_adj}()}.
For further info, see their docstrings.
}
\keyword{internal}
