% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsearch.R
\name{helper_search}
\alias{helper_search}
\title{Helper function for gtsearch}
\usage{
helper_search(series, is_keyword, timeframe, as_tbl_ts = TRUE)
}
\arguments{
\item{series}{A dataframe from \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}} with Google data.}

\item{is_keyword}{A logical argument to determine, if we searched beforehand
for a keyword (TRUE) or a category (FALSE).}

\item{timeframe}{A time frame to search the queries in consisting of
a start date and an end date in YYYY-MM-DD form.}

\item{as_tbl_ts}{Logical value to determine if tibble should already
be coerced to a tsibble with keyword or category column as key.}
}
\description{
This is a helper function for \code{\link[=gtsearch]{gtsearch()}} to
catch some errors and to mutate the Google series
to our liking
}
\keyword{internal}
