% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifwtrends.R
\docType{package}
\name{ifwtrends}
\alias{ifwtrends}
\title{ifwtrends: A package for evaluating data from Google Trends}
\description{
The package ifwtrends is based on trendecon and GtrendsR and takes
data from Google Trends to support economic forecasting.
}
\section{Functions}{

\code{\link{pca}()}

\code{\link{roll}()}

\code{\link{daily_series}()}

\code{\link{factorR2}()}

For in-depth information and showcases, see the vignette
at \code{vignette(topic = "ifwtrends-demo", package = "ifwtrends")}
}

