% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Gibt pca fuer Backtesting zurueck}
\usage{
roll(
  keyword = NA,
  category = 0,
  geo = "DE",
  start_series = "2006-01-01",
  start_period = "2014-01-01",
  end = Sys.Date(),
  fun = trendecon::ts_gtrends,
  ...
)
}
\arguments{
\item{keyword}{Eine character-Vektor mit dem Suchbegriffen}

\item{category}{Ein Numeric Vektor mit den Kategorien}

\item{geo}{Die Region}

\item{start_series}{Das Startdatum der Zeitreihen.}

\item{start_period}{Das Startdatum des Ausgabefensters.}

\item{end}{Das Enddatum der Zeitreihen.}

\item{fun}{Funktion, die auf die rollende Zeitreihe angewendet werden soll.}

\item{...}{Zusaetzliche Parameter, die an die Funktion in fun weitergegeben werden}
}
\value{
Monatliche Tabelle mit pca in jeder Spalte. Je Spalte wird ein neuer Monat hinzugenommen.
}
\description{
\code{roll} Gibt fuer start_period bis end die jeweils dann aktuelle Berechnung von pca aus.
}
\examples{
\dontrun{
roll(keyword = c("ikea", "saturn"), start_period = "2018-01-01", end = "2020-01-01")
}
}
