% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_daily_series.R
\name{simple_daily_series}
\alias{simple_daily_series}
\title{Generate a daily time series from a broader time frame.}
\usage{
simple_daily_series(
  keyword = "arbeitslos",
  geo = "DE",
  from = "2006-01-01",
  verbose = TRUE
)
}
\arguments{
\item{keyword}{A string with a single search query. At the moment,
multiple search queries are not allowed.}

\item{geo}{A geographical region to restrict the search query to.}

\item{from}{Start date of the search query.}

\item{verbose}{Boolean value. True, if some status messages
should be displayed in the console during runtime.}
}
\value{
A tibble with a date column and a value column of a
daily time series.
}
\description{
(experimental) This is a more simplistic approach to generate daily
data in a much broader time frame, where Google usually
doesn't provide daily data at all. See more in section
Method.
}
\section{Method}{

This function searches for multiple time periods
with a duration of 9 months with sometimes significantly
overlapping ranges of Google Trends data. The overlapping
period is then used to ensure that the time series has
a consistent scale. As of now, \code{simple_daily_series}
and \code{daily_series} have a somewhat comparable
trend to a certain degree. Important to note,
\code{simple_daily_series} won't have a value in the time
series that is above 100. But this function will also have outliers.
If you search for a smoothed (and adjusted) time series, try
the \code{\link[tempdisagg:td]{tempdisagg::td()}} function from the \link\link{tempdisagg} package.
}

\examples{
simple_daily_series <- function(keyword = "covid-19",
geo = "DE",
from = "2020-01-01",
verbose = TRUE)

}
