% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gttrend_adj.R
\name{trend_adj}
\alias{trend_adj}
\title{Trend adjustment}
\usage{
trend_adj(series, method = "moving_avg", log.trafo = FALSE)
}
\arguments{
\item{series}{The input tibble in tidy form
with columns \code{time}, \code{value} and optional column \code{id}.
Be careful, that when using the method \code{moving_avg}, you can only
use a series with one keyword respective category.}

\item{method}{Character which method for
trend adjustment should be choosen. See Details.}

\item{log.trafo}{Logical, indicates if value
should be transformed to log(value).}
}
\value{
Returns a tibble with trend adjusted values and a
date column.

For trend_method there can be choosen between
\code{"firstdiff"}, \code{"moving_avg"} and \code{"comtrend"}.
If you choose \code{"firstdiff"},
first differences with \code{lag = 1} is executed.
If you choose the moving average,
the time series will be decomposed into its components
and the trend will be subtracted
from the whole time series (using loess).
With \code{"comtrend"},
there is a polynom of degree 5
with id-fixed effects estimated,
which captures the common trend.
The residuals where then used as the adjusted series.
For further detail, see Woloszko et al. (2020). Attention:
This method is not implemented as of the current development state.
}
\description{
\code{trend_adj} takes a tibble
of time series and returns a tibble with trend adjusted values.
}
\examples{
series <- trendecon::ts_gtrends("ikea", time = "all")
trend_adj(series, log.trafo = TRUE, method = "moving_avg")
}
\keyword{internal}
