% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_series.R
\name{daily_series}
\alias{daily_series}
\title{Konsistente taegliche Zeitreihe}
\usage{
daily_series(keyword = c("arbeitslos"), geo = "DE", from = "2006-01-01")
}
\arguments{
\item{keyword}{Der Suchbegriff. Bis jetzt nur einer möglich}

\item{geo}{Region}

\item{from}{Startdatum

Vorsicht, braucht sehr lange und verursacht viele Google-Suchanfragen.}
}
\value{
Tabelle der taeglichen Werten
}
\description{
\code{daily_series}. Schaetzt mit Chow-Lin eine konsistente lange Zeitreihe mit taeglicher Frequenz zurueck.
}
\examples{
\dontrun{
daily_series(keyword = "Ikea", geo = "NL", from = "2021-01-01")
}
}
