% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_frequencies.R
\name{daily_series}
\alias{daily_series}
\title{Konsistente taegliche Zeitreihe
\code(daily_series) Schaetzt mit Chow-Lin eine lange taegliche Reihe.}
\usage{
daily_series(keyword = c("arbeitslos"), geo = "DE", from = "2006-01-01")
}
\arguments{
\item{keyword}{Der Suchbegriff. Bis jetzt nur einer möglich}

\item{geo}{Region}

\item{from}{Startdatum}
}
\value{
Tabelle der taeglichen Werten
}
\description{
Konsistente taegliche Zeitreihe
\code(daily_series) Schaetzt mit Chow-Lin eine lange taegliche Reihe.
}
\examples{
daily_series(keyword = "Ikea", geo = "NL", from = "2008-01-01")
}
