% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trend.R
\name{est_trend}
\alias{est_trend}
\title{Estimates the trend of Google search queries}
\usage{
est_trend()
}
\value{
Returns a tibble with a
time series, which consists of a polynomial trend
of degree five of a category.
}
\description{
\code{est_trend} downloads for a sample
of 250 random Google Trends categories the relative search volume.
Based on this, a common trend is calculated and saved in an
internal dataframe the user cannot access
but the packages' functions do.
However, the user can use this
function to calculate the trend for himself.
The internal dataframe used in other functions is updated on a
monthly basis (on the seventh day of each month) on GitHub.
}
\section{Warning}{

This function takes a few minutes to work and generates
a lot of traffic at Google. An IP ban is therefore quite likely.
}

\examples{
\dontrun{
est_trend()
}
}
