% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trend.R
\name{est_trend}
\alias{est_trend}
\title{Schätzt den Trend einer GTrends-Kategorie}
\usage{
est_trend()
}
\value{
Gibt eine Zeitreihe zurück, die den polynomialen Trend
fünften Grades einer Kategorie enthält.
}
\description{
\code{est_trends}. Schätzt den Trend einer GTrends-Kategorie

Schätzt den gemeinsamen Trend aller Google Reihen.
}
\examples{
\dontrun{
est_trends()
}
}
