% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorR2.R
\name{factorR2}
\alias{factorR2}
\title{R squared of the series' regression on factors}
\usage{
factorR2(series, factors, plot = F)
}
\arguments{
\item{series}{A tibble with multiple time series as columns.}

\item{factors}{A tibble with factors as columns.}

\item{plot}{A logical indicator whether the function should
additionally return a bar plot of the \eqn{R^2}.}
}
\value{
A tibble of \eqn{R^2} from each time series for each factor will be
returned.
If \code{plot = TRUE}, a plot containing the \eqn{R^2} is returned
and displayed additonally.
}
\description{
\code{factorR2} returns the \eqn{R^2}
of the series' regression on the factors
in form of a tibble.
}
\examples{
dat <- pca(
  keywords = c("Pluto", "Saturn"),
  categories = 0,
  geo = "DE",
  time = paste("2020-01-01", "2020-06-01")
)

series <- dplyr::select(dat, date, 4:5)
factors <- dplyr::select(dat, date, 2:3)

factorR2(series, factors, plot = TRUE)
}
