% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factorloadings.R
\name{factorR2}
\alias{factorR2}
\title{R2 der Regression der Serien auf die Faktoren
\code(factorR2)}
\usage{
factorR2(series, factors)
}
\arguments{
\item{series}{tibble mit den Zeitreihen als Spalten.}

\item{factors}{tibble mit den Faktoren als Spalten.}
}
\value{
Tabelle der R^2 jeder Zeitreihe auf jeden Faktor
}
\description{
R2 der Regression der Serien auf die Faktoren
\code(factorR2)
}
\examples{
2+2
}
