% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_m.R
\name{forecast_m}
\alias{forecast_m}
\title{Monthly forecast evaluation of vintages}
\usage{
forecast_m(r_list, dat, fd = T)
}
\arguments{
\item{r_list}{todo}

\item{dat}{todo}

\item{fd}{todo}
}
\value{
Todo
}
\description{
Todo
}
\examples{
1 + 1
1 + 2
}
