% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_q.R
\name{forecast_q}
\alias{forecast_q}
\title{Quarterly forecast evaluation of vintages}
\usage{
forecast_q(r_list, dat, fd = T)
}
\arguments{
\item{r_list}{A Google Trends time series in form of a tibble (or tsibble)
which is used to predict the next quarter in data.}

\item{fd}{A logical value to indicate if you want to use the first differences
in r_list (true) or normal values (false).}

\item{data}{A tibble with some data
you want to predict for the next quarter
that is somewhat correlated with the data in r_list.}
}
\value{
Returns a list with the the predicted data and the used model
results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=forecast_q]{forecast_q()}} uses quarterly vintages generated by \code{\link[=roll]{roll()}}
to predict the value of another time series for the next quarter
Due to complications, this function \strong{does not work currently}.
}
\section{Methods}{

Google Data should be log-transformed
currently it uses a RIDGE-Regression
but other models such as OLS, LASSO or
PCA are available.
}

\examples{
1+1
1 + 2
}
