% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_q.R
\name{forecast_q}
\alias{forecast_q}
\title{Quarterly forecast evaluation of vintages}
\usage{
forecast_q(r_list, dat, fd = T)
}
\arguments{
\item{r_list}{todo}

\item{dat}{todo}

\item{fd}{todo}
}
\value{
Todo
}
\description{
Todo
}
\examples{
1+1
1 + 2
}
