% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_index.R
\name{g_index}
\alias{g_index}
\title{Preparation of Google Trends data}
\usage{
g_index(
  keyword = NA,
  category = 0,
  geo = "DE",
  time = str_c("2006-01-01 ", Sys.Date()),
  lags = 0
)
}
\arguments{
\item{keyword}{A vector (chr) with search requests.}

\item{category}{A vector (num) with category numbers from Google Trends.}

\item{geo}{A geographical region to restrict the search queries to.}

\item{time}{Time period from where the relative values should be taken
(for more information, visit the
documentation of \code{\link[trendecon]{ts_gtrends}}).}

\item{lags}{Number of delays in additional columns (max. value: 4).}
}
\value{
Firstly, each row will be log transformed and
seasonal adjusted (via RJDemetra's X-13 ARIMA methods).
Furthermore, the first derivatives of these adjusted time
series will be returned (optionally with additional columns containing
lags).
}
\description{
\code{g_index} downloads for
various search requests respective trends categories
data and applies a logarithmic transformation aswell as a
seasonal adjustment.
The function returns first derivatives (lagged if desired).
}
\examples{
\dontrun{
g_index(keyword = c("ikea", "saturn"), time = "2018-01-01 2021-01-01")
}
}
