% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsearch.R
\name{gtsearch}
\alias{gtsearch}
\title{Simple Google Trends data search}
\usage{
gtsearch(
  keywords = NA,
  categories = 0,
  geo = "DE",
  timeframe = paste("2006-01-01", Sys.Date())
)
}
\arguments{
\item{keywords}{A vector (chr) of keywords to search for.}

\item{categories}{A vector (numeric) of category numbers to search for.}

\item{geo}{The region to search in.}

\item{timeframe}{A time frame to search the queries in.}
}
\value{
A tibble with a time series of Google Trends search volume from
given inputs.
}
\description{
\code{gtsearch} is a simple wrapper
around the \code{gtrends} function from package. It only
returns a time series in form of a tibble with the actual
(in relative terms) search volume of keywords or categories.
}
\details{
\code{gtsearch} returns a simple tibble with the original
interest data in a keyword or category from Google Trends.
}
\examples{
gtsearch(keywords = c("pluto", "saturn"), timeframe = "2020-01-01 2020-06-01")
}
