% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsearch.R
\name{gtsearch}
\alias{gtsearch}
\title{Simple Google Trends data search}
\usage{
gtsearch(
  keyword = NA,
  category = NA,
  geo = "DE",
  timeframe = paste("2006-01-01", Sys.Date()),
  as_tbl_ts = TRUE
)
}
\arguments{
\item{keyword}{A vector (chr) of keywords to search for.}

\item{category}{A single numeric value stating the category number.
To search for multiple categories at once is as of today not possible.
Use an \code{for}-loop instead.}

\item{geo}{The region to search in (e.g. for Germany "DE").}

\item{timeframe}{A time frame to search the queries in consisting of
a start date and an end date in YYYY-MM-DD form.}

\item{as_tbl_ts}{Logical value to determine if tibble should already
be coerced to a tsibble with keyword or category column as key.}
}
\value{
A tsibble with a time series of Google Trends search volume from
given inputs and the keyword respective category as key for it.
}
\description{
\code{gtsearch} is a simple wrapper
around the \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}} function. It only
returns a time series in form of a tibble with the actual
(in relative terms) search volume of keywords or categories.
}
\details{
\code{gtsearch} returns a simple tibble with the original
interest data in a keyword or category from Google Trends.
}
\examples{
gtsearch(keywords = c("pluto", "saturn"), timeframe = "2020-01-01 2020-06-01")
}
