% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifwtrends.R
\docType{package}
\name{ifwtrends}
\alias{ifwtrends}
\title{ifwtrends: A package for evaluating data from Google Trends}
\description{
The package ifwtrends is based on trendecon and gtrendsR and takes
data from Google Trends to support economic forecasting.
}
\section{Functions}{

\itemize{
\item \code{\link[=daily_series]{daily_series()}}: Creates a daily series for time frames where Google
usually doesn't provide daily data. Uses the Chow-Lin-method.
\item \code{\link[=est_trend]{est_trend()}}: Estimates a common trend between all Google Trends categories
and uses that to trend adjust in some other functions. This is updated on
a monthly basis in the package. However, the user may use this function
for himself on his own.
\item \code{\link[=factorR2]{factorR2()}}: Estimates the R squared between keywords.
\item \code{\link[=forecast_m]{forecast_m()}}: Makes a monthly forecast on a time series based on the data
of some Google Trends data. \strong{Does currently not work correctly.}
\item \code{\link[=forecast_q]{forecast_q()}}: Makes a quarterly orecast on a time series based on the data
of some Google Trends data. \strong{Does currently not work correctly.}
\item \code{\link[=gtpreparation]{gtpreparation()}}: Makes a search query and applies a
trend adjustment (with the common
trend), a seasonal adjustment and, if wished, adds lag variables
to the resulting data frame.
\item \code{\link[=gtsearch]{gtsearch()}}: Makes a simple Google search for either
a keyword or a category.
\item \code{\link[=gtseas_adj]{gtseas_adj()}}: Seasonal adjusts a given time series or uses
\code{\link[=gtsearch]{gtsearch()}} internally to seasonal
adjust a search query. The seasonal adjustment
is done via X-13-ARIMA-SEATS from \code{\link[seasonal:seas]{seasonal::seas()}}.
\item \code{\link[=gttrend_adj]{gttrend_adj()}}: Trend adjusts a given time series or uses
\code{\link[=gtsearch]{gtsearch()}} internally to trend
adjust a search query. It either can use
first differences, a moving average or the common trend computed
by \code{\link[=est_trend]{est_trend()}} to adjust the time series.
\item \code{\link[=pca]{pca()}}: Does a Principal Component Analysis on a search query.
\item \code{\link[=roll]{roll()}}: Creates some rolling Google
search queries for different time frames.
\item \code{\link[=simple_daily_series]{simple_daily_series()}}: A more simplistic alternativ for
\code{\link[=daily_series]{daily_series()}} that doesn't rely on inner functions from other packages.
}

For in-depth information and showcases, see the German vignette
at \code{vignette(topic = "ifwtrends-demo", package = "ifwtrends")}
}

