% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Hauptkomponentenanalyse}
\usage{
pca(
  keywords = NA,
  categories = 0,
  geo = "DE",
  start = "2006-01-01 CET",
  end = Sys.Date()
)
}
\arguments{
\item{keywords}{Eine character-Vektor mit dem Suchbegriffen}

\item{categories}{Ein Numeric Vektor mit den Kategorien}

\item{geo}{Die Region}

\item{start}{Das Startdatum der Zeitreihen.}

\item{end}{Das Enddatum der Zeitreihen.}
}
\value{
Monatliche Tabelle der Hauptkomponenten und der Zeitreihen.
}
\description{
\code{pca} Berechnet fuer mehrere Suchbegriffe oder mehrere Kategorien die Hauptkomponenten der monatlichen Zeitreihen
}
\examples{
pca(keywords = c("ikea", "saturn"),start = "2018-01-01", end = "2020-01-01")
}
