% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Principal component analysis on Google Trends time series}
\usage{
pca(
  keyword = NA,
  category = 0,
  geo = "DE",
  time = str_c("2006-01-01 ", Sys.Date())
)
}
\arguments{
\item{geo}{A geographical region to restrict the search queries to.}

\item{keywords}{A vector (chr) with search queries (or a single search query).}

\item{categories}{A vector (num) with Google Trends category numbers.}

\item{start}{Time series start date.}

\item{end}{Time series end date.}
}
\value{
Tibble with monthly principal components next to the actual time series.
}
\description{
\code{pca} computes for several search queries
or several categories the principal
components of the monthly time series.
}
\examples{
\dontrun{
pca(keywords = c("ikea", "saturn"), start = "2018-01-01", end = "2020-01-01")
}
}
