% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Returns a tibble for backtesting with PCA.}
\usage{
roll(
  keyword = NA,
  category = 0,
  geo = "DE",
  start_series = "2006-01-01",
  start_period = "2014-01-01",
  end = Sys.Date(),
  fun = trendecon::ts_gtrends,
  ...
)
}
\arguments{
\item{keyword}{A vector (chr) with search queries (or a single search query).}

\item{category}{A vector (num) with Google Trends category numbers.}

\item{geo}{A geographical region to restrict the search queries to.}

\item{start_series}{Start date of the time series.}

\item{start_period}{Start date of the returned time frame.}

\item{end}{End date of time series and to be returned time frame.}

\item{fun}{Name of a function, that will be applied to a time series.}

\item{...}{Additional parameter for
the function that is in \code{fun} specified.}
}
\value{
Tibble with monthly data where \code{\link[=pca]{pca()}} is
applied on every column.
For each column, a new month is added.
}
\description{
\code{roll}
Returns for the time from start_period to end
the respective result from \code{\link[=pca]{pca()}} back.
}
\examples{
roll(keyword = "ikea", start_series = "2018-01-01", start_period = "2019-01-01", end = "2020-01-01")
}
