% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{roll}
\alias{roll}
\title{Gibt}
\usage{
roll(
  keywords = NA,
  categories = 0,
  start_series = "2006-01-01",
  start_period = "2014-01-01",
  end = Sys.Date(),
  components = max(length(keywords), length(categories))
)
}
\description{
Gibt
}
