% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_adjust.R
\name{seas_adj}
\alias{seas_adj}
\title{Seasonal adjustment}
\usage{
seas_adj(series, freq = "month", log.trafo = F, method = "arima")
}
\arguments{
\item{series}{The input tibble in tidy
form with columns \code{time},
\code{value} and optional column \code{id}.
Monthly or quarterly frequency.}

\item{freq}{Character "month" or "quarter" for the frequency.}

\item{log.trafo}{Logical,
indicates if value should be transformed to log(value).}

\item{method}{Character, which method for
adjustment should be choosen. See Details.}
}
\value{
Returns a tibble with trend adjusted values and a
date column.

For method, there can be choosen \code{"firstdiff"} and \code{"arima"}.
If \code{"firstdiff"}, first derivatives with \code{lag = 1} are computed.
If \code{"arima"}, the X-13ARIMA-SEATS  procedure is used
(performed by the \code{\link[seasonal:seas]{seasonal::seas()}} function from the seasonal package).
}
\description{
\code{seas_adj} takes a tibble of
time series and returns a tibble with seasonal adjusted values.
}
\examples{
series <- trendecon::ts_gtrends(c("ikea", "saturn"), time = "all")
seas_adj(series, freq = "month", log.traf = TRUE, method = "firstdiff")
}
