% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_adjust.R
\name{trend_adj}
\alias{trend_adj}
\title{Trend adjustment}
\usage{
trend_adj(series, log.trafo = F, method = "firstdiff")
}
\arguments{
\item{series}{The input tibble in tidy form with columns \code{time}, \code{value} and optional column \code{id}.}

\item{log.trafo}{Logical, indicates if value should be transformed to log(value).}

\item{method.}{Character which method for trend adjustment should be choosen. See Details.

For trend_method there can be choosen \code{"firstdiff"} and \code{"comtrend"}.
If \code{"firstdiff"}, first differences with \code{lag = 1} is executed.
If \code{"comtrend"}, there is a polynom of degree 5 with id-Fixed Effects estimated, which captures the common trend. The residuals where then used as the adjusted series. For further Detial see Woloszko et.al. (2020)}
}
\description{
\code{trend_adj} Takes a tibble of timeseries and returns tibble with trend adjusted values.
}
\examples{
series <- ts_gtrends(c("ikea", "saturn"), time = "all")
trend_adj(series, log.trafo = T, method = "firstdiff")
}
